fix(migrations): align Vitess foreign key types

+++ database/migrations/vitess/0000000004-create-board_columns-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `board_columns` (
   `id` bigint PRIMARY KEY auto_increment,
-  `board_id` integer not null REFERENCES `boards`(`id`),
+  `board_id` bigint not null REFERENCES `boards`(`id`),
   `name` varchar(80) not null,
   `position` integer,
   `card_limit` integer,
